requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.2
openai>=1.40.0
werkzeug==3.0.1
